move tag list stuff into its own concern
