+++ app/controllers/creators_controller.rb
@@ -1,5 +1,6 @@
 class CreatorsController < ApplicationController
   include Filterable
+  include TagListable
 
   before_action :get_creator, except: [:index, :new, :create]
 
@@ -11,7 +12,6 @@ class CreatorsController < ApplicationController
     end
 
     @tags, @unrelated_tag_count = generate_tag_list(@filters.empty? ? nil : @models, @filter_tags)
-    @tags = @tags.select(:name, :taggings_count)
 
     # Ordering
     @creators = case session["order"]
